fix(tests): Update imports after client_selector_widget rename
Fix test imports to use client_settings_dialog module after renaming.
Resolves CI/CD test collection error.
